+++ PetMatchmakerWeb/src/app/app.routes.ts
@@ -6,7 +6,7 @@ import { SignUpComponent } from './login-components/sign-up/sign-up.component';
 import { AdminLoginComponent } from './login-components/admin-login/admin-login.component';
 import { QuestionnaireComponent } from './questionnaire/questionnaire.component';
 import { PetLisitingComponent } from './pet-lisiting/pet-lisiting.component';
-import { AdminDashboardComponent } from './admin-dashboard.component';
+import { AdminDashboardComponent } from './admin/admin-dashboard.component';
 
 export const routes: Routes = [
   { path: 'welcome', component: WelcomeComponent },
